Pull out kmc core from counter and remove using std from header files
As per many recommendations, using using namespace std in header files
is not recommended (e.g.
http://stackoverflow.com/questions/4872373/why-is-including-using-namespace-into-a-header-file-a-bad-idea-in-c)

I also noticed that the kmer counter was different than much of the
code it relies on. I pulled out the core and separated it from the
counter to make it a little more sturctured.